Dr Quine Added
The project cannot be corrected until Transcendance is valided...
But this requierds itself to have teammates who are available and ready
to work on it. So yea, will take a few to get it there.

Otherwise, Dr Quine is a project about self-replication, revolving
around software that output its own source code.